Move old checkbox under deprecated
