+++ packages/components/Checkbox/src/deprecated/Checkbox.settings.android.ts
