+++ packages/components/Checkbox/src/deprecated/Checkbox.settings.ios.ts
