+++ packages/components/Checkbox/src/deprecated/Checkbox.settings.macos.ts
