+++ packages/components/Checkbox/src/deprecated/Checkbox.settings.ts
